requirements.txt
pyyaml>=6.0
jsonschema>=4.0.0
requests>=2.28.0
schedule>=1.2.0